refactor: 홈 화면 UI 구조 개선 및 분석 FAB 오버레이 구현
* **refactor: `HomeScreen` 컴포저블 분리 및 구조화**
    * `HomeScreen.kt`의 거대한 UI 로직을 `HomeScreenContent`와 `HomeAnalysisFabOverlay`로 분리하여 가독성과 유지보수성을 높였습니다.
    * 홈 화면의 상태(Loading, Empty, Content)에 따른 레이아웃 처리 로직을 `HomeScreenContent`로 통합했습니다.
    * `HomeScreen` 프리뷰 코드에서 `CompositionLocalProvider`를 통해 필요한 상태를 주입하도록 개선했습니다.

* **feat: `HomeAnalysisFabOverlay` 컴포넌트 추가**
    * 음성 녹음 분석 및 파일 업로드 분석 기능을 제공하는 확장형 플로팅 메뉴를 구현했습니다.
    * `PopupPositionProvider`를 사용하여 메뉴의 확장/축소 상태에 따른 동적 팝업 위치 계산 로직을 적용했습니다.
    * `LocalAppDimmerState`와 연동하여 메뉴 확장 시 배경 딤 처리가 적용되도록 구현했습니다.

* **refactor: `PrezelApp` 레이아웃 및 네비게이션 로직 정리**
    * `PrezelApp`의 UI 구조를 `AppNavigationContent`와 `AppDimmerOverlay` 컴포저블로 분리하여 역할을 명확히 했습니다.
    * 반복되는 네비게이션 애니메이션 설정을 `defaultPrezelNavTransition` 함수로 공통화했습니다.

* **refactor: 데이터 소스 및 UI 모델 정리**
    * `PresentationRemoteDataSourceImpl`에서 파일 처리를 위한 헬퍼 함수들을 파일 수준의 프라이빗 함수로 이동하여 클래스 내부 로직을 단순화했습니다.
    * `GrowthGraphData` 모델에서 사용되지 않는 `selectedItem` 프로퍼티를 제거했습니다.

* **fix: UI 텍스트 및 프리뷰 데이터 수정**
    * `PresentationSheet`의 D-Day 표기 형식을 `-3`에서 `D-3`으로 수정했습니다.
    * `HomeScreen` 프리뷰에서 그래프 데이터가 올바르게 표시되도록 샘플 데이터를 보강했습니다.
