+++ Prezel/app/src/main/java/com/team/prezel/ui/PrezelApp.kt
@@ -1,5 +1,6 @@
 package com.team.prezel.ui
 
+import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.SharedTransitionLayout
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
@@ -24,6 +25,7 @@ import com.team.prezel.core.common.event.GlobalEvent
 import com.team.prezel.core.common.event.GlobalEventBus
 import com.team.prezel.core.designsystem.component.PrezelNavigationScaffold
 import com.team.prezel.core.designsystem.component.PrezelNavigationScope
+import com.team.prezel.core.designsystem.theme.PrezelTheme
 import com.team.prezel.core.navigation.LocalNavigator
 import com.team.prezel.core.navigation.Navigator
 import com.team.prezel.core.navigation.ProvideSharedTransitionScope
@@ -78,48 +80,64 @@ private fun PrezelAppContent(
     Box(modifier = Modifier.fillMaxSize()) {
         SharedTransitionLayout {
             ProvideSharedTransitionScope(this@SharedTransitionLayout) {
-                val provider = remember(entryBuilders, navigator) {
-                    entryProvider {
-                        entryBuilders.forEach { builder -> this.builder() }
-                    }
-                }
-
-                PrezelNavigationScaffold(
-                    showNavigationBar = appState.shouldShowNavigationBar,
-                    snackbarHostState = LocalSnackbarHostState.current,
-                    navigationItems = { AppNavigationItems(appState = appState, navigateToKey = { key -> navigator.navigate(key) }) },
-                ) { padding ->
-                    NavDisplay(
-                        entries = appState.navigationState.toEntries(provider),
-                        onBack = navigator::goBack,
-                        modifier = Modifier.padding(padding),
-                        transitionSpec = {
-                            fadeIn(animationSpec = tween(durationMillis = 100)) togetherWith
-                                fadeOut(animationSpec = tween(durationMillis = 100))
-                        },
-                        popTransitionSpec = {
-                            fadeIn(animationSpec = tween(durationMillis = 100)) togetherWith
-                                fadeOut(animationSpec = tween(durationMillis = 100))
-                        },
-                        predictivePopTransitionSpec = {
-                            fadeIn(animationSpec = tween(durationMillis = 100)) togetherWith
-                                fadeOut(animationSpec = tween(durationMillis = 100))
-                        },
-                    )
-                }
+                AppNavigationContent(
+                    appState = appState,
+                    entryBuilders = entryBuilders,
+                    navigator = navigator,
+                )
             }
         }
 
-        if (appDimmerState.isVisible) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(com.team.prezel.core.designsystem.theme.PrezelTheme.colors.scrimContainer)
-                    .noRippleClickable { appDimmerState.dismiss() },
-            )
-        }
+        AppDimmerOverlay(isVisible = appDimmerState.isVisible, onDismiss = appDimmerState::dismiss)
     }
 }
+
+@Composable
+private fun AppNavigationContent(
+    appState: PrezelAppState,
+    entryBuilders: ImmutableSet<EntryProviderScope<NavKey>.() -> Unit>,
+    navigator: Navigator,
+) {
+    val provider = remember(entryBuilders, navigator) {
+        entryProvider {
+            entryBuilders.forEach { builder -> this.builder() }
+        }
+    }
+
+    PrezelNavigationScaffold(
+        showNavigationBar = appState.shouldShowNavigationBar,
+        snackbarHostState = LocalSnackbarHostState.current,
+        navigationItems = { AppNavigationItems(appState = appState, navigateToKey = navigator::navigate) },
+    ) { padding ->
+        NavDisplay(
+            entries = appState.navigationState.toEntries(provider),
+            onBack = navigator::goBack,
+            modifier = Modifier.padding(padding),
+            transitionSpec = { defaultPrezelNavTransition() },
+            popTransitionSpec = { defaultPrezelNavTransition() },
+            predictivePopTransitionSpec = { _: Int -> defaultPrezelNavTransition() },
+        )
+    }
+}
+
+@Composable
+private fun AppDimmerOverlay(
+    isVisible: Boolean,
+    onDismiss: () -> Unit,
+) {
+    if (!isVisible) return
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(PrezelTheme.colors.scrimContainer)
+            .noRippleClickable(onClick = onDismiss),
+    )
+}
+
+private fun defaultPrezelNavTransition(): ContentTransform =
+    fadeIn(animationSpec = tween(durationMillis = 100)) togetherWith
+        fadeOut(animationSpec = tween(durationMillis = 100))
 
 @Composable
 private fun ObserveGlobalEvents(
